Add LabVIEW code to replicate the Mpemba effect
This code contains LabVIEW files and methods required to replicate the Mpemba effect from Kumar and Bechhoefer (2020). This is a slightly modified version of Avinash Kumar's code -- with some functionality to adapt it to the FPGA currently used in the Bechhoefer lab, some bug fixes, slightly better documentation, and some added methods.